GENZWEAR WEBSITE
This is my First Semester Final Project of Introduction to Information and Communication Technology Lab. It is a simple website which is made up of HTML and CSS, a little bit of javascript was also used. The Website consists of 6 webpages (Home, About, Shop, Contact, Cart, Product). It is only a frontend, no backend is used. And It have only limited Features. I would live to hear some suggestions about future upgrades from you guys. The Report, Proposal, and a presentation are also attached in the folder.